Remove GeminiClient class and refactor dependencies
